+++ src/app.rs
@@ -158,15 +158,15 @@ impl Application for PulseStreamApp {
                 });
 
                 self.save_settings();
-                self.streamer.start(
+                self.streamer.start(crate::audio::StreamConfig {
                     server,
                     port,
                     rate,
                     channels,
                     device_id,
                     process_id,
-                    self.state.mute_local_output,
-                );
+                    mute_local_output: self.state.mute_local_output,
+                });
             }
 
             Message::Disconnect => {